changed int to float

+++ algebra/addition/c/add.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 
 int main(){
-  int num1, num2;
+  float num1, num2;
 
   printf("Enter the first number:\n");
-  scanf("%d", &num1);
+  scanf("%f", &num1);
   printf("Enter the second number:\n");
-  scanf("%d", &num2);
-  printf("%d + %d = %d\n", num1, num2, num1+num2);
+  scanf("%f", &num2);
+  printf("%.2f + %.2f = %.2f\n", num1, num2, num1+num2);
   return 0;
 }
